redaction_service/requirements.txt
fastapi
uvicorn
python-consul
fpdf
pydantic
requests
spacy
pypdf
python-multipart
sqlalchemy

pytest
httpx

pytesseract
pdf2image
pillow
